backends/mlx_backend/requirements.txt
mlx>=0.22.0
tokenizers>=0.15.0
Pillow>=9.0.0
huggingface_hub>=0.20.0
numpy>=1.24.0
